@@ -1,7 +1,7 @@
 package dev.gtuk.diga
 
 import dev.gtuk.diga.dtos.BillingRequest
-import javax.validation.Valid
+import jakarta.validation.Valid
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.GetMapping
