@@ -1,8 +1,8 @@
 package dev.gtuk.diga.dtos
 
+import jakarta.validation.constraints.NotEmpty
+import jakarta.validation.constraints.Size
 import java.util.Date
-import javax.validation.constraints.NotEmpty
-import javax.validation.constraints.Size
 
 data class BillingRequest(
     @get:NotEmpty @get:Size(min = 16, max = 16) val code: String,
